requirements.txt
PyYAML==6.0
openpyxl==3.1.2
tqdm==4.65.0
pyOpenSSL==23.2.0
cryptography==41.0.1